first test

+++ src/main/scala/model/ImagesInfo.scala
@@ -1,6 +1,6 @@
 package model
 
 case class ImagesInfo(
-  // Number of images in database
-  count: Int)
+  // Number of images
+  count: Long)
